+++ Arrays1D/src/com/avi/com/avi/devlopeapps/MicroAndArrayUpdate.java
@@ -1,4 +1,4 @@
-package com.devlopeapps;
+package com.avi.devlopeapps;
 import java.util.Arrays;
 import java.util.Scanner;
 
@@ -11,7 +11,7 @@ public class MicroAndArrayUpdate {
         for (int i = 0; i < t; i++) {
             int n=sc.nextInt();
             int k=sc.nextInt();
-            int arr[]=new int[n];
+            int[] arr =new int[n];
             for(int j=0;j<n;j++)
                 arr[j]=sc.nextInt();
             Arrays.sort(arr);
